Create README.md
Support your local Phi Theta Kappa Chapter, Alpha Delta Theta.
Phi Theta Kappa is the honor society for 2-year college students. Its purpose is to not only reward academic excellence, but to foster growth and provide fellowship among its members. Alpha Delta Theta, the Downtown Jefferson CTC Chapter, is dedicated to providing scholarship, leadership, service, and fellowship opportunities for its members; the four hallmarks for its members; the four hallmarks of Phi Theta Kappa.

About Alpha Delta Theta:
- Alpha Delta Theta is currently a 4 star chapter, and has plans to become a 5 star chapter by 2015. This requires a lot of time and dedication from its members.
- Part of the 5 star chapter plan is to research and write about an Honors in Action topic. The topic is meant to benefit the college, the community, and beyond. It will also require a service project and a college project, both meant to be of benefit and significance.
- We provide the resources for unique and innovative programs that encourage, engage, and motivate individuals to complete their degrees, fulfill their dreams and become leaders in our society. We have a pledge to complete college program in place, and encourage students to use collegefish.org to ease their transfer into a 4-year college.

Alpha Delta Theta - Phi Theta Kappa is able to play a vital role in the lives of extraordinary students due to the generosity of our philanthropic partners who support the Society's mission through outright gifts, sponsorships, and in-kind donations. Your investment enables Alpha Delta Theta to expand our volunteer programs (we are currently partnering with DoSomething.org; collecting old clothing to be reused and/or recycled) and we just finished walking in the March for Babies, sponsored by the March of Dimes. It also allows us to attract talented speakers that aid in developing our leadership programs. It allows us to travel to other schools in our region, which develops fellowship, and scholarship in our members through the lively exchange of ideas. It also enables us to focus on and implement our Honors in Action study topic, which includes our service and college projects that are put in place to serve the community, the college and beyond. It is through your generous donations that we may focus on our hallmarks of leadership, service, fellowship, and scholarship.

We thank you for your support!